Fixing issue with annotations.
With the newest android SDK the template was failing when trying to
parse some of the LINT files left in dex files. Unsure if this was
always happening or if this was the first time I'd ever seen it.
Either way the fix I found was to change the variables to local ones
which will cause the template to function properly.